:recycle: api: Move session-for-attendance-sheet injection to session
