+++ api/src/certification/session/application/attendance-sheet-controller.js
@@ -1,4 +1,4 @@
-import { usecases } from '../../shared/domain/usecases/index.js';
+import { usecases } from '../domain/usecases/index.js';
 
 const getAttendanceSheet = async function (request, h) {
   const sessionId = request.params.id;
