+++ api/tests/certification/session/unit/application/attendance-sheet-controller_test.js
@@ -1,5 +1,5 @@
 import { attendanceSheetController } from '../../../../../src/certification/session/application/attendance-sheet-controller.js';
-import { usecases } from '../../../../../src/certification/shared/domain/usecases/index.js';
+import { usecases } from '../../../../../src/certification/session/domain/usecases/index.js';
 import { expect, hFake, sinon } from '../../../../test-helper.js';
 import { getI18n } from '../../../../tooling/i18n/i18n.js';
